+++ lib/hpke/src/aead_cipher.cpp
@@ -214,8 +214,7 @@ AEADCipher::open(const bytes& key,
   }
 
   auto inner_ct_size = ct.size() - tag_size;
-  // NOLINTNEXTLINE (cppcoreguidelines-narrowing-conversion)
-  auto tag = bytes(ct.begin() + inner_ct_size, ct.end());
+  auto tag = bytes(ct.begin() + static_cast<long>(inner_ct_size), ct.end());
   if (1 != EVP_CIPHER_CTX_ctrl(ctx.get(),
                                EVP_CTRL_GCM_SET_TAG,
                                static_cast<int>(tag_size),
